requirements.txt
botcity-framework-web
botcity-maestro-sdk
pandas
requests
openpyxl
smtplib
email